+++ tapo/src/requests.rs
@@ -7,12 +7,14 @@ mod get_trigger_logs;
 mod handshake;
 mod login_device;
 mod multiple_request;
+mod play_alarm;
 mod secure_passthrough;
 mod set_device_info;
 mod tapo_request;
 
 pub use crate::responses::TemperatureUnitKE100;
 pub use energy_data_interval::*;
+pub use play_alarm::*;
 pub use set_device_info::*;
 
 pub(crate) use control_child::*;