+++ CHANGELOG.md
@@ -8,6 +8,10 @@ file. This change log follows the conventions of
 
 ### Rust
 
+#### Added
+
+- Added functionality for controlling the alarm on the H100 hub via the `play_alarm` and `stop_alarm` methods in the `H100Handler`. Additionally, `get_supported_ringtone_list` is available to retrieve the list of supported ringtones for debugging purposes. (thanks to @kay)
+
 ### Python
 
 ## [v0.8.0][v0.8.0] - 2024-12-07